[Mod] DBM Echo: Compute echo coefficients as floats.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@5922 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/plugins/DigiBoosterEcho.h
@@ -57,9 +57,9 @@ protected:
 	uint32 m_sampleRate;
 
 	// Echo calculation coefficients
-	int32 m_PMix, m_NMix;
-	int32 m_PCrossPBack, m_PCrossNBack;
-	int32 m_NCrossPBack, m_NCrossNBack;
+	float m_PMix, m_NMix;
+	float m_PCrossPBack, m_PCrossNBack;
+	float m_NCrossPBack, m_NCrossNBack;
 
 	// Settings chunk for file I/O
 	PluginChunk chunk;
